requirements.txt
openai
requiests
typer
langchain
gigachat
python-dotenv

